pyproject.toml
[project]
name = "de-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "dbt-postgres>=1.10.0",
    "duckdb>=1.5.0",
    "pandas>=3.0.1",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=23.0.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "pytest>=9.0.2",
    "ruff>=0.15.6",
]


[tool.ruff]
indent-width = 4
